requirements.txt
bert-serving-client==1.10.0
bert-serving-server==1.10.0
pandas==1.0.1
Flask==1.1.1
scikit-learn==0.22.1
spacy==2.2.3
requests==2.23.0
numpy==1.18.1
matplotlib==3.2.0
seaborn==0.10.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
spacy-stanfordnlp==0.1.3
sentence-transformers==0.2.5.1

